[Autofill] Extract FormFieldData::max_length of textareas
Not only text-control <input> elements but also <textarea> supports
the maxlength attribute. This CL sets FormFieldData::max_length also
for textareas.

This CL changes the default maxiumm length for text-control <input>
and <textarea> elements to the maximum 32 bit number. Before this
CL, it was the maximum int, whose size isn't specified.

The reason that this CL chooses the maximum 32 bit number and not
64 bit, is to avoid the risk of integer overflows in the, for
example, the rationalization code.

Therefore, this CL moves the maxlength default value from
WebInputElement to form_autofill_util.cc.

Bug: 1484810
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4874062
Commit-Queue: Christoph Schwering <schwering@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1205150}

+++ components/autofill/core/browser/form_parsing/credit_card_field.cc
@@ -35,7 +35,7 @@ namespace {
 
 // Returns true if a field that has |max_length| can fit the data for a field of
 // |type|.
-bool FieldCanFitDataForFieldType(int max_length, ServerFieldType type) {
+bool FieldCanFitDataForFieldType(uint64_t max_length, ServerFieldType type) {
   if (max_length == 0)
     return true;
 
@@ -602,7 +602,7 @@ bool CreditCardField::ParseExpirationDate(AutofillScanner* scanner,
   scanner->RewindTo(month_year_saved_cursor);
 
   // Bail out if the field cannot fit a 2-digit year expiration date.
-  const int current_field_max_length = scanner->Cursor()->max_length;
+  const uint64_t current_field_max_length = scanner->Cursor()->max_length;
   if (!FieldCanFitDataForFieldType(current_field_max_length,
                                    CREDIT_CARD_EXP_DATE_2_DIGIT_YEAR))
     return false;
