+++ components/password_manager/core/browser/password_generation_frame_helper.cc
@@ -126,7 +126,7 @@ std::u16string PasswordGenerationFrameHelper::GeneratePassword(
     const GURL& last_committed_url,
     autofill::FormSignature form_signature,
     autofill::FieldSignature field_signature,
-    uint32_t max_length) {
+    uint64_t max_length) {
   autofill::PasswordRequirementsSpec spec;
 
   // Lookup password requirements.
