+++ components/password_manager/core/browser/password_generation_frame_helper.h
@@ -76,7 +76,7 @@ class PasswordGenerationFrameHelper {
       const GURL& last_committed_url,
       autofill::FormSignature form_signature,
       autofill::FieldSignature field_signature,
-      uint32_t max_length);
+      uint64_t max_length);
 
  private:
   friend class PasswordGenerationFrameHelperTest;
